refac(vc, icons): use getters and setters for easier augmentation
The version control segment is now controlled by a single update
function that gets both state and info and then sets a local variable.
This function is still called after the same hooks.

The getter for the info is now advised instead in the icons code.

The main package can now deal with the segment being a function or a
variable.