feat: wal-line-create-static-segment
Adds new macro to unify creation of static segments. It accepts a
getter to get the segment and a setup and teardown lambda. Internally
it creates functions to get and set the segment so they can be
advised.

Uses the macro in icons, project, vc and buffer-name (which is now its
own package).

Adds const wal-line--feature-like that is appended to
wal-line-features on minor-mode setup so we can load buffer-name.

The render function now checks if the segment is in initial state in
which case the setter is called, otherwise it returns the segment as
is. This allows an initial loading that is not hook-based.