feat: macro for dynamic segments
Adds equivalent macro for creating dynamic segments and uses it for
org and mc segments.

This also adapts the rendering so that the priority can be set to
`'current` to only show a segment for the current window.

This also re-defines the whale segment as a static segment.

This improves the render function to evaluate to the setter or the
variable before returning the `:eval` list.