reat: better setups and teardowns
Segment setups and teardowns are now easier to reason about. Segments
are only set up or torn down during `whale-line-{setup,teardown}-hook`
if the segment is part of `whale-line-segments`. When segments are
rebuild, only the added/removed segments are set up/torn down.
Augments are handled differently, they **need** a verify function that
is called when the hooks are run. During rebuild, they are not
setup/torn down (the same goes for segments that use `:verify`).